refactor: refactored api calls to module. (#112)
